chore: rename locale.domain
