--- frontend/core/stores/locale.domain/index.ts
@@ -1,22 +1,0 @@
-import { proxy } from 'valtio'
-import { LOCALE } from '~/const/i18n'
-import type { TLocale } from '~/spec'
-
-import type { TInit, TStore } from './spec'
-
-export default ({ locale = LOCALE.EN, localeData = '{}' }: TInit): TStore => {
-  const store = proxy({
-    locale,
-    localeData,
-
-    // actions
-    setLocale: (locale: TLocale): void => {
-      store.locale = locale
-    },
-    setLocaleData: (localeStr: string): void => {
-      store.localeData = localeStr
-    },
-  })
-
-  return store
-}
