--- frontend/core/stores/locale.domain/spec.d.ts
@@ -1,11 +1,0 @@
-import type { TLocale } from '~/spec'
-
-export type TInit = {
-  locale: TLocale
-  localeData: string
-}
-export type TStore = TInit & {
-  // actions
-  setLocale: (locale: TLocale) => void
-  setLocaleData: (localeStr: string) => void
-}
