+++ frontend/core/stores/locale/spec.d.ts
@@ -1,9 +1,10 @@
 import type { TLocale } from '~/spec'
 
-export type TStore = {
+export type TInit = {
   locale: TLocale
   localeData: string
-
+}
+export type TStore = TInit & {
   // actions
   setLocale: (locale: TLocale) => void
   setLocaleData: (localeStr: string) => void
