+++ frontend/landing/app/layout.tsx
@@ -4,7 +4,7 @@ import type { Metadata } from 'next'
 import { Suspense } from 'react'
 import { LANDING_COMMUNITY } from '~/const/name'
 import GlobalLayout from '~/providers/GlobalLayout'
-import LocaleStoreProvider from '~/stores/locale.domain/provider'
+import LocaleStoreProvider from '~/stores/locale/provider'
 import { CommunityInfoProvider } from '~/stores/provider.landing'
 import { ssrThemeInitScript } from '~/utils/ssr/script'
 import Main from './Main'
